Pass the absolute path of package being tested to revgrep
Some tools provide an absolute path to the file name, but the patches
used by revgrep have relative paths. Revgrep uses the current working
directory by default to make absolute paths relative, but because the
tools are being ran inside a different path than revgrep, it's not
possible to match issue filename with patch filename.

Instead, we'll pass in the path to the package to revgrep. We haven't
had to do this before because golint, go vet and apicompat were the
only tools in use, and all their paths are relative.