Clone to import path in /home/bradleyf/go/src/
Currently we clone into $GOPATH/src/gopherci but any project that
imports sub packages in the same repository, would have those
dependencies go get'd (go got?).

If the project imported internal packages, those couldn't be built
at all with the error `use of internal package not allowed`.

This fix attempts to guess the import path, based on the URL by:

- Getting the base repo's HTML URL `https://github.com/owner/repo`
- Strippping the scheme `github.com/owner/repo`
- Docker or file system executer runs command with that as the
working directory
- All clones, tools etc, are then executed in the repositories
working directory, which should be $GOPATH/src/github.com/owner/repo

I'm not a big fan of this is done (just stripping the scheme), but
the only edge case I can forsee are those with vanity import paths,
and it's not clear how to handle them right now, potentially
automated or potentially it could be configuration option per repo.

The other edgecases will come if GitHub changes URL structure or
we support other providers where the import path is not clear, both
of which can be dealt with at the time.

Fixes #16.