Add timeout to GCPPubSub connection setup
There's been a couple cases now (#46) when creating the topic on
GCP Pub/Sub that has caused GopherCI to block and never start.

This change creates a context with a timeout for use by the
various pub sub sdk's setup methods with a fixed (hardcoded)
timeout. A test is also provided which sets a very low timeout and
ensures a context.DeadLineExceeded is received.

Unfortunately, a leaktest needed to be removed from an adjacent
test, which caused false positives. I did not have time to correctly
debug this situation.

Fixes #46.