Support synchronize and reopened events on GitHub Pull Requests
Previously we were only checking a PR when it was first opened,
and not when additional commits were pushed.

There was no reason we were ignoring it, so this commit supports
that behaviour now.

We'll also build on reopened events, see #24 for some discussion
as there's room to optimise this workflow, but I get one main
benefit that it makes development faster by just having to close
and open a PR to have it rebuilt.

Resolves #24.
Resolves #25.