Support GitHub PushEvent
This is accomplished with a few refactors.

We no longer fetch the diff URL provided in a Pull Request event,
it was convenient, but it's not available in Pushes and we can
easily calculate it ourselves without the need for the extra
lookup. This was GitHub specific too, other providers may not have
provided that URL.

Analyse does need to process a Push differently than a Pull Request,
and this is managed by the EventType type.

We ignore the GitHub push event's Before SHA ref, during testing
force pushes, this would be the ref of the commit that's being
"overwritten" so it wouldn't exist when clonsed. Unlike a
non-forced push, where the commit exists after the clone. So instead
of using Before, we use After~len(Commits), so no matter how many
commits are being pushed, or whether the previous commit exists or
not, we should be able to process it.

Resolves #27.