Write analysis to the database
This provides the foundations for a web view (discussed in #28),
so users can view the issues without relying on inline or issue
comments.

This also provides a mechanism to time how long the analysis is
taking as well as a per tool timing metric. This allows us to better
understand how much time different repositories or installations spend
checking.

We record which tools were executed, even if no issues were detected
so the user knows which checks passed. This will cause the analysis_tool
table to become very large over time.

Relates to #3.