Add integration tests
This is the first attempt at running some basic integration tests.

Right now, the tests focus on PRs only and ensures a comment is left.
It includes a small help IntegrationTest to start GopherCI and some
helper functions, such as waiting for status to change from pending
to success.

Currently, these are not automatically executed in CI.