Fix two slice out of bounds bugs when deduplicating comments
The first would occur because we were manipulating the slice in place
and when there's multiple issues, a slice out of bounds or the wrong
issue could be removed. Fixed by starting at the end of slice and
working to the start.

The second issue could occur when there are already duplicate
comments, if an issue matches an exiting comment, remove it then stop
checking for more duplicate comments for that issue. Removing it again
could remove the wrong issue or cause a slice out of bounds exception.