Ignore issues in generated code, or vendored/testdata directories
We consider generated code, anything that appears to be generated
or is in a vendor/testdata directory. GopherCI uses
github.com/gopherci/isFileGenerated which mostly uses a function
provided by github.com/shurcooL/go.

In using isFileGenerated, it makes it difficult for a user to provide
additional files to be ignored, this would need to be implemented in
similar areas by other method if we decide to support that option.

This causes all issues from generated code to be ignored, there
will be no option to only comment, or only set the build to failed
as the issue will be ignored as soon as it's detected.

An additional integration test has been provided, because although
there are unit tests, it's possible the underlying tools misbehave
or change their implementation details.

Resolves #57.