Working web UI
GopherCI itself has the ability to show the results of an analysis.
This is the first version, very rough cut.

It's achieved by creating an internal package web, containing static
assets, templates and the handlers for the web views. This is done
to best compartmentalise the web components from the CI checks as
much as possible, as GopherCI is a CI tool, not a web view into the
results. I don't imagine too much more functionality being added
that isn't directly related to viewing analysis results.

Additional changes were required, there was a bug when writing results
to the database, it was using a TIME(3) field but trying to store
seconds in it, so results above 60 failed. The new Duration field
resolves this, and cleans up the marshalling of dates in and out of the
database. AnalysisStatus also implements sql.Scanner, so it can easily
be read from the database.

db.StartAnalysis now returns a db.Analysis with zero values zero for
all members except ID which is set correctly. Analyser now accepts
a pointer to a db.Analysis and it will populate the members as required.
This was done to allow the initial github SetStatus call to set the
HTML URL of the analysis. So when the status API is set the first time
a user can view and access the results.

Included is some changes to the GitHub CallbackHandler, it seems to be
deprecated, but because we had it set initially (I never used it),
GitHub still redirects users to this URL to then be redirected to the
analysis page. A small change was made to continue handling those
redirects and avoiding being an open redirect as well.

The web template itself is a very very rough draft, more work will
be made, but initially I just wanted a basic working version so lay
good foundations to focus on the UI at a later stage. It looks terrible,
but we'll address that later.