View a diff of the changes in the analysis view
We'll do this by storing before and after commits and pr number in
the analysis table. They'll be used for display, but also to generate API
endpoints instead of storing the actual API URLs.

https://api.github.com/repositories/{repository_id}/pulls/{request_number}
https://api.github.com/repositories/{repository_id}/compare/{commit_from}...{commit_after}

GitLab has similar URLs which only requires merge request number or from
and to SHAs.

We've also created a new interface called VCSViewer, this will be used
to abstract away the VCS, such as GitHub, GitLab and others, providing
a uniform interface. It will likely be expanded upon in the furture to
return other information. github.Installation will now begin to implement
this interface.

The recently added package "web" currently cares for this interface, I
don't believe this will be its final resting place, web has no business
handling this interface when it's likely we'll add an email package at
some stage in the future. So something will be moved or renamed, I just
don't know what.