Add 3 issues for staticcheck, golint and unused (to then be reverted)
GopherCI needs some example analysis o be used in URL monitoring as well
as general examples. Using a third party repo (such as the one in the
integration tests repo) risks the reports failing when the diffs are
no longer available.

For this reason, I thought it might just be best for this repo to
have the examples, as this should always exist (GopherCI handles
repositories being renamed) and there won't be any force pushes.

This commit will then be reverted.