commitFrom should be relative to commitTo for GitHub
Previously, for GitHub, we used the before sha, for a new branch
this is 0000000000000000000000000000000000000000 which the API will
not accept when comparing commits https://developer.github.com/v3/repos/commits/#compare-two-commits

Further, when force pushing the before commit no longer exists,
which also causes the compare commit API to return errors. Instead,
we'll use commitTo~<number of commits> which works for both cases.
This is the same method used when analysing the request as well.