Handle repositories with no history (eg new)
For a new repository, or a branch which doesn't have any history
(tracking a different git tree), we can't get a diff using the
"git diff abcd~5...abcd" method.

Instead, we'll try that way, and if it fails, try to get the diff
using "git show abcd".

I chose not to have the analyser check the ref "abcd~5" exists first
to help simplify the code, with less conditionals, but I'm not
convinced this is any more readable and potentially less reliable.
Therefore, it may be refactored in the future.

I've added an integration test for this as well, to ensure the
behaviour with git is always tested and not mocked like it is in the
unit test.

Fixes #79.