Ignore GitHub events early, such as no Go repos and no installation
This commit sets up a better framework for ignoring certain GitHub
events early, before they're sent to the queue.

We don't queue them because some organisations have many non-Go
repositories and queuing all those jobs to just be immediately
discarded was becoming a little wasteful. In the future, I'm happy
for this decision to be reversed.

This is implemented in a GitHub specific manner, as it ensured the
events would be discarded before a clone of the repository occurred.

Resolves #80.
Related to #73.