Support reading per repository configurations from .gopherci.yml
Users should be able to apply per repository configuration for GopherCI via
a config file in the repository, this allows non-admin users to modify
GopherCI on a per repository basis if they have write access. Which practically
means users can also have their pull request modify GopherCI if they require
additional packages to be installed (for cgo for example) or if they want to
modify the tools being run.

We achieve this by including a YAMLConfig which implements a configReader
interface, an interface was chosen primarily to keep tests simpler, but
the analyser doesn't really care where the config came from, just that it's
able to receive one.

The YAMLConfig should be preloaded with the global settings and then it should
filter the config based on reading the repository's config file, and finally
returning the config to analyser to use.

Relates #8.