Analysis page fetching diff should not be a critical error
The db.Analysis type contains a list of the issues per tool that were
stored in the database. We use these in two ways:

A) Provide it to web.DiffIssues function to combine a diff/patch with
the issue to show the issue inline.

B) Show a summary of the issues, grouped by tool.

If fetching the diff/patch fails, we can still show the issues per
tool, this contains all the information a user really requires.

Now, patches is only set in the template if there were issues *and*
there was no errors fetching the diff.

Relates to #88.