Override docker's semver detection v17.05.0-ce is stable
dep init determines docker (moby) uses semver due to its syntax
and it was using valid semver syntax until they switched from
v1.13.1 to v17.03.0-ce.

dep assumes the -ce suffix is a pre-release and therefore the latest
stable version is v1.13.1 and not v17.05.0-ce.

This change overrides dep's choice to be the latest stable, but will
require us to manually change this if we require later versions.