Analysis to accept an Executer instead of creating its own
Previously I was trying to keep as much logic out of the GitHub
specific handler as I could that wasn't directly GitHub related.
So the analyser creating its own execution environment was OK
initially. But it's added some knowledge the Analyser doesn't
require, and eventually the analyser will be refactored more into
its own parts the the Analyse method may just run a single tool
and a constructor will prepare the environment for the tools.

Also, we'll soon be wanting to wrap/decorate the Executer to have
the database log the output of all the commands being issued.

So the GitHub handler, that already knew about the analyser keeps
that knowledge to itself and just passes a working executor and
manages the clean up itself.