db.DB to provide an analysis.Executer for recording executions
DB must now implement a mechanism to record (or discard) results
of an Execution. This is because the user operating GopherCI and
the user viewing the results of a CI analysis may be different
people with different access, when an error occurrs, previously
only the operator would see the error, not the user who may be
able to fix it.

We need a mechanism for user's to view the results of an Execution,
including the output. This is done by requiring the database to
provide an analysis.Executer that the analysis can then wrap/decorate
a real executer, to record the output.

*SQLDB also has a Cleanup method which runs background tasks, such
as removing old outputs that are attached to an analysis that's older
than 30 days. This is a crude cleanup method, in the future we may
wish to have this configurable, as different values for different
statuses, such as failures keeping their records for longer to
assist in debugging.