Ignore GitHub private repositories
We do not currently support private repositories because we do not
use tokens to clone the repositories URLs, and the build status
pages are not currently authenticated.

This results in clone errors failing the build, and if that is fixed
could result in information leak on the build summary page.

As a stop gap, we'll ignore requests for private repositories. This
won't provide any feedback to the user, which is unfortunate, but to
do that would require a different way to ignore repositories that
provides feedback, the longer term goal is to simply support private
repositories correctly.

Fixes #73.
Relates to #22.