Show environment information in analysis output
Summary page now shows the output of all process run, it'll be
helpful for users to see:

- Go version: we change this on behalf of the user as newer versions
are released, and may allow users to chose a version later.
- Go environment: we may need to adjust some environment parameters
to support different versions and other options that the user may
need to be aware of.
- Limits: we limit the maximum virtual memory a process can use and
other limits are in place, users should be able to see these limits.
- OS Release: users have the ability to install apt packages, as a
result they will need to at least know the distribution version to
know the package names.

We output the versions after (attempting) to read the config file so
future versions that change the environment (such as different Go
versions) will have the status reflect their configuration.