Summary page to only fetch diff if there are issues to show
Summary page fetches the diff from GitHub in order to show the issues
inline with a generated patch.

Previously we were fetching the diff even if there were no issues to
show. This change ensures the diff is fetched only when there's at
least one issue.

This optimisation can shave off ~400ms time to first byte, when showing
an analysis with no issues.