Integration tests to test for states being set in order
Previously the github.Analyse test was responsible for testing that
the Status API states were being set in the correct order. This
complicated the testing and we'll shortly be duplicating some of that
functionality and tests into its own type.

In moving to its own type, we won't be able to track the order of
statuses, we want to see Pending be set first, followed by Success. The
unit tests will test to make sure the status API can set the status
correctly, but won't have visibility in the order they occur.

Instead, the integration tests, which are already checking the status
API statuses will be responsible to ensure the state transitions from
Pending to Success/Failure or Error.