Comment on commits via CommitCommentReporter and InlineCommitCommentReporter
Previously we were only commenting on pull requests, this adds support for
commenting on commits (eg during a Push Event).

Unlike pull requests, where inline comments are positioned in relation to
the entire pull request (no matter how many commits there are), commit
comments must be on the individual commit and positioned relative to that
commit. The comments aren't on the push, they're on the individual commits.
making it very difficult to track which issue was caused by which commit.

GopherCI runs after applying all commits, the diffs it generates are the
accumulation of all commits, and the individual line number of issues are
after all commits are applied. Making it very difficult to track which issue
was caused by which commit.  See full discussion in #75 for why other
workarounds wouldn't work.

Instead, if the push event only has one commit, we can place comments
inline and we use InlineCommitCommentReporter for this.

If there are multiple commits, we'll use CommitCommentReporter to create
a single comment on the most recent commit with a summary saying:

    GopherCI found %d issues in the last %d commits, see: https://...

The user can then visit the URL to see all issues and they also appear
inline on the analysis summary page.

When commit comments are combined with pull requests, the following
behaviour occurs (tested at the time of writing):

If a single comment was created with CommitCommentReporter, the pull
request will show this single comment on the main pull request page,
then the pull request event will create inline comments for the same
issues. So you'll see n+1 comments, in different locations on the
GitHub UI.

If multiple inline comments were created with InlineCommitCommentReporter
all inline comments are shown on the main pull request page (they don't
appear inline on the pull request), and then the pull request event
will create additional inline comments for the same issues. So you'll
see n*2 comments, in different locations on the GitHub UI.

If this behaviour is not desired, there may be some options:

- Disable push event event analysis.
- Disable pull request event analysis.
- Pull request to delete existing commit comments.
- Pull request to check for existing commit comment and skip.
- There may be other options too.

I'm wary about deleting commit comments, there may have been some
discussion there.

Resolves #75.