GitHub add PRReviewReporter to post create Pull Request Reviews
PRReviewReporter uses GitHub's Pull Request Reviews to post issues,
there are multiple benefits, the main being multiple issues result
in just one email to the user.

Additionally, when no issues are detected, a review is still created
with the event APPROVE. This will become configurable at a later
stage as it results in an extra email.

Whilst possible to set the review event to REQUEST_CHANGES, this
hasn't been enabled yet, but will become a configuration option at
a later stage. Instead, the event is set to COMMENT if there are issues
found.