GitHub to post reviews using PRReviewReporter by default
For pull requests, this enables the Pull Request Review reporter to
approve pull requests or post comments if issues are found.

The unit test for GitHub has been simplified further to just check
if a comment has been made. In the future the github package won't
require this mocking at all as it will just provide reports to an
analyse function, and instead rely on the integration tests to ensure
issues are reported.

An additional integration test has been included to verify a review
is posted when there are no issues, and its state is APPROVED.

Resolves #112.