Provide application wide logger for structured, levelled with context
./internal/logger provides a single logger than all packages can use
to provide:

Levelled logs: Debug is for developers, Info for general production
messages to assist in troubleshooting, Error to be used when an error
occurs that's handled but needs to be raised to an operator and Fatal
which logs the message and terminates execution.

Structured: Using With() to add fields to a message allows operators to
easily view related messages, as GopherCI provides many concurrent
services, it's useful to only inspect message related to a certain
event.

Context: Using With() to return another Logger which logs all messages
with the provided name/key and value. Allows developers to pass along
loggers with information about the event so new clients can be unaware
of the event but still have additional context added to their logs.

Logger is its own interface that wraps a single type of logger. Currently
it is logrus, chosen because it matches requires, I'm familiar with it,
and supports hooks.

Sentry is also used for the hooks, where any event of level Error or
above will send the event to a Sentry instance, where operators can
collate, deduplicate be alerted and track resolutions.

Finally, due to case sensitive import conflicts of github.com/sirupsen/logrus
one of the copies is moved to the GOPATH to avoid conflicts with the other.
This is very unfortunate hack until everyone moves to lowercase the
import path.