Skip GitHub pull request events that aren't accessible
I've had a few cases where a user or repository that is not private
is not accessible. It appears to be because GitHub has flagged the
account due to suspicious behaviour.

This means we receive the hooks, but can't access some of the required
API endpoints, such as /repos/owner/repo/pulls/1.

So, this change adds an extra check for all PRs to check if the
pulls API responds with an error, then we'll log and ignore the event
completely.

See https://platform.github.community/t/pull-request-event-not-setting-private-field/2981.
See https://sentry.io/bradley-falzon/gopherci/issues/329097380/.